Implement Minters, Splitters and Burners

+++ src/TokenSystem/Exceptions/NonPositiveTokenAmountException.cs
@@ -1,12 +1,13 @@
 // Copyright (c) Comrade Coop. All rights reserved.
 
 using System;
+using System.Numerics;
 
 namespace TokenSystem.Exceptions
 {
 	public class NonPositiveTokenAmountException : Exception
 	{
-		public NonPositiveTokenAmountException(decimal tokenAmount)
+		public NonPositiveTokenAmountException(BigInteger tokenAmount)
 			: base($"Non-positive token amount requested: {tokenAmount}.")
 		{
 		}
