+++ src/TokenSystem/Tokens/IReadOnlyTaggedTokens.cs
@@ -1,13 +1,14 @@
 // Copyright (c) Comrade Coop. All rights reserved.
 
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace TokenSystem.Tokens
 {
-	public interface IReadOnlyTaggedTokens<TTagType> : IEnumerable<KeyValuePair<TTagType, decimal>>
+	public interface IReadOnlyTaggedTokens<TTagType> : IEnumerable<KeyValuePair<TTagType, BigInteger>>
 	{
-		decimal TotalTokens { get; }
+		BigInteger TotalTokens { get; }
 
-		decimal GetAmountByTag(TTagType tag);
+		BigInteger GetAmountByTag(TTagType tag);
 	}
 }
